Cupy support

+++ hyperspy/_signals/hologram_image.py
@@ -104,7 +104,7 @@ def _parse_sb_size(s, reference, sb_position, sb_size, parallel):
                              'neither reference nor hologram dimensions.')
         # sb_position navdim=0, therefore map function should not iterate:
         else:
-            sb_size_temp = np.float64(sb_size.data)
+            sb_size_temp = float(sb_size.data)
     else:
         sb_size_temp = sb_size.deepcopy()
     return sb_size, sb_size_temp
@@ -462,7 +462,7 @@ class HologramImage(Signal2D):
             # sb_position navdim=0, therefore map function should not iterate
             # it:
             else:
-                sb_smoothness_temp = np.float64(sb_smoothness.data)
+                sb_smoothness_temp = float(sb_smoothness.data)
         else:
             sb_smoothness_temp = sb_smoothness.deepcopy()
 
@@ -561,7 +561,7 @@ class HologramImage(Signal2D):
             if reference.axes_manager.navigation_size == 0 and \
                sb_smoothness.axes_manager.navigation_size > 0:
                 # 1d reference, but parameters are multidimensional
-                sb_smoothness_ref = np.float64(
+                sb_smoothness_ref = float(
                     _first_nav_pixel_data(sb_smoothness_temp))
             else:
                 sb_smoothness_ref = sb_smoothness_temp
